modelBackend/requirements.txt
pandas
django-cors-headers
joblib>=0.14
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18